+++ src/business/services/deposit/deposit.service.ts
@@ -1,6 +1,6 @@
 import { Injectable } from '@nestjs/common';
 import { DepositEntity, DepositRepository } from 'src/data';
-import { DepositDto } from 'src/presentation/dtos';
+import { DepositDto } from 'src/data/dtos';
 import { DataRangeModel, PaginationModel } from 'src/data/models';
 
 @Injectable()
