+++ src/presentation/index.ts
@@ -1,2 +1,2 @@
 export * from './controllers';
-export * from './dtos';
+export * from '../data/dtos';
